一阶滤波改为浮点数接口

+++ tzbox.h
@@ -48,6 +48,6 @@ uint64_t TZBoxNtohll(uint64_t n);
 
 // TZBoxFilterFirstOrder 一阶滤波
 // value是当前值.realValue是实时值.ratio是系数,从0开始,系数越大滤波效果越好
-int TZBoxFilterFirstOrder(int value, int realValue, int ratio);
+float TZBoxFilterFirstOrder(float value, float realValue, float ratio);
 
 #endif
